test: Fixes PR' comments
